feat: Added a ping endpoint that returns a json object
Added a ping end point to get some experience with routing and sending
http requests to servers. The button that is used is unstylized because
it will shortly get removed in future updates to this little project.
Also note that the url is pointing to `localhost`. don't want to keep
that in there for obvious reasons