refactor: Port the UI to use lustre-ui
Didn't really like learning CSS all that much so I decided to port the
whole thing to a component library that takes care of most of the
styling for me. All that really needs to get done is picking of a colour
theme and voila! Pretty decent looking front end application.

Removed the top bar until there's something to actually display there.
The anchor code and such isn't quite the same as vanilla lustre so I'm
gonna figure it out later